[feat] Handler & Validator 수정
- Handler & Validator 수정

+++ src/main/kotlin/toy/webflux/develop/validator/MemberCreateValidator.kt
@@ -13,6 +13,6 @@ class MemberCreateValidator : Validator {
         val memberCreate: MemberCreate = target as MemberCreate
 
         if(memberCreate.name.isBlank()) errors.rejectValue("name", "field.empty", "The name must not be empty.")
-        if(memberCreate.age < 0) errors.rejectValue("age", "field.invalid", "The age must be positive number")
+        if(memberCreate.age.toInt() < 0) errors.rejectValue("age", "field.invalid", "The age must be positive number")
     }
 }
